Bug #51316 up: Fix temporary file handling
Use modern tempfile.TemoraryDirectory as context manager instead of
tempfile.mkdtemp() and atexit() as the later does only performs the
cleanup when the Python interpreter is shutting down *cleanly*.

It does not happen on os._exit() and not on os.execve(), which the
updater uses internally. This leads to /tmp/ being filled up, especially
during unit testing.