Bug #53461: If bind via LDAPI fails retry after 1s ten times before aborting
Workaround for unfortunate interplay between systemd and U@S import:

> 4. why doesn't systemd start the service again

There is a 10 second timeout between restart notifier restarts configured
in the service definition. If the service is explicitly stopped during
that time the restart cycle is aborted until the service is explicitly
started again.

In this case a ucs@school import test was running. The ucs@school import
stops the notifier service but only starts it again if it was already
running, which it wasn't because it had just crashed.
(keyword: "stopped_notifier")